add better debug logging...
...which showed me why the video was painfully slow on the onboard_ui.  The rendering async function was hogging all of the async io loop with just an asyncio.sleep(0).  clock.tick from py game is not async aware and not actually yielding any of the loop.

which is to say that we now have full frame rate, low latency video streaming from the webapp (browser) back to onboard_ui and being displayed.

There is still more.  We need to center crop the frames (in handle_video_frames) to 1080x1080 and possibly add scaling.  I also want to expiriment with locking on to a face (the biggest one) in the frame and displaying that in the 1080x1080 full screen display.