another abandoned claude branch...
there is a bunch of additional chaff in this set of changes and for this branch as a whole that I am want to abandon this branch. This is the result of overconfident debugging by Claude which was sure it was fixing problems even though the outcome was exactly the same :/

there were some diamonds tho.
- this attempt (unlike the one abandoned on bee/webRTCStreaming branch) did not interfere with video renderingfps
- the audio is audible and very low latency but it is garbled
- as part of its overwhelming want to add diagnostic code, Claude added a means of saving the frames for the first few seconds to a wav file
- I then asked it to write a debug program debug/test-sdplay that used sounddevice and existing libs to play a wav file.  the wav file from the previous diamond played well through the debug program without all the cruff currently baked in to audio_renderer.py on this branch.

Next step is to start another branch off main and give it the test-sdplay.py as the only starting point.