fix issue where performance is webRTC performance is degraded after switching between MJPEG and WebRTC in the UI
I was today years old when I discovered that removing an image from the dom, where the image is sourced to a multipart response that never dies, causes Chrome to continue streaming the MJPEG stream.  Changing the src of the image is the only way I've found to cause Chrome to abort the connection.